fix: dont enable until verification succeeds
