+++ internal/user/service.go
@@ -7,7 +7,6 @@ import (
 	"context"
 
 	"github.com/autobrr/autobrr/internal/domain"
-	"github.com/autobrr/autobrr/pkg/errors"
 )
 
 type Service interface {
@@ -15,10 +14,11 @@ type Service interface {
 	FindByUsername(ctx context.Context, username string) (*domain.User, error)
 	CreateUser(ctx context.Context, req domain.CreateUserRequest) error
 	Update(ctx context.Context, req domain.UpdateUserRequest) error
+	Delete(ctx context.Context, username string) error
 	Enable2FA(ctx context.Context, username string, secret string) error
-	Verify2FA(ctx context.Context, username string, code string) error
-	Disable2FA(ctx context.Context, username string) error
+	Store2FASecret(ctx context.Context, username string, secret string) error
 	Get2FASecret(ctx context.Context, username string) (string, error)
+	Disable2FA(ctx context.Context, username string) error
 }
 
 type service struct {
@@ -36,24 +36,10 @@ func (s *service) GetUserCount(ctx context.Context) (int, error) {
 }
 
 func (s *service) FindByUsername(ctx context.Context, username string) (*domain.User, error) {
-	user, err := s.repo.FindByUsername(ctx, username)
-	if err != nil {
-		return nil, err
-	}
-
-	return user, nil
+	return s.repo.FindByUsername(ctx, username)
 }
 
 func (s *service) CreateUser(ctx context.Context, req domain.CreateUserRequest) error {
-	userCount, err := s.repo.GetUserCount(ctx)
-	if err != nil {
-		return err
-	}
-
-	if userCount > 0 {
-		return errors.New("only 1 user account is supported at the moment")
-	}
-
 	return s.repo.Store(ctx, req)
 }
 
@@ -61,18 +47,22 @@ func (s *service) Update(ctx context.Context, req domain.UpdateUserRequest) erro
 	return s.repo.Update(ctx, req)
 }
 
+func (s *service) Delete(ctx context.Context, username string) error {
+	return s.repo.Delete(ctx, username)
+}
+
 func (s *service) Enable2FA(ctx context.Context, username string, secret string) error {
 	return s.repo.Enable2FA(ctx, username, secret)
 }
 
-func (s *service) Verify2FA(ctx context.Context, username string, code string) error {
-	return s.repo.Verify2FA(ctx, username, code)
-}
-
-func (s *service) Disable2FA(ctx context.Context, username string) error {
-	return s.repo.Disable2FA(ctx, username)
+func (s *service) Store2FASecret(ctx context.Context, username string, secret string) error {
+	return s.repo.Store2FASecret(ctx, username, secret)
 }
 
 func (s *service) Get2FASecret(ctx context.Context, username string) (string, error) {
 	return s.repo.Get2FASecret(ctx, username)
+}
+
+func (s *service) Disable2FA(ctx context.Context, username string) error {
+	return s.repo.Disable2FA(ctx, username)
 }
